+++ tests/Unit/FactoryTest.php
@@ -3,7 +3,6 @@
 use DomainValidity\Factory;
 use DomainValidity\Validator;
 
-
 test(Factory::class . '::make produce instance of ' . Validator::class, function () {
     $validator = getInstance();